refactored chunking
