+++ extraction/src/utils/mod.rs
@@ -1,5 +1,6 @@
-pub mod db;
 pub mod configs;
+pub mod db;
+pub mod json2mkd;
 pub mod rrq;
 pub mod server;
 pub mod storage_service;
